sched/tune: Introduce SchedTune Assist[v3]
This implements a mechanism by which default SchedTune parameters
can be configured in-kernel, circumventing userspace, and
mitigating reliance on ramdisk modification in the context of
custom kernels.

[2.5V]: This version adds proper protection
from userspace (mainly init) trying to write lame
boost values and gives full control to developer
and user (sh is not blocked).

[V3.0]: Use a struct to store all the values.

[0ctobot: Update for msm-4.9 and improve coding style]
[yaro: Update for 4.14]
Co-authored-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
[wight554: strip for 4.4 and update tunables for msm8998]
Signed-off-by: Volodymyr Zhdanov <wight554@gmail.com>

Adapt to Kernel 4.19
Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>